Reshuffled printer, small fixes
